Matrix addition + refactoring
